edit libs and utils
