+++ libs/models/module/__init__.py
@@ -1,0 +1,1 @@
+from .non_local import NONLocalBlock3D
