+++ libs/spatial_transforms.py
@@ -1,0 +1,280 @@
+"""
+Copyright (c) 2017 Kensho Hara
+Released under the MIT license
+https://github.com/kenshohara/3D-ResNets-PyTorch
+"""
+
+import random
+
+from torchvision.transforms import transforms
+from torchvision.transforms import functional as F
+from PIL import Image
+
+
+class Compose(transforms.Compose):
+
+    def randomize_parameters(self):
+        for t in self.transforms:
+            t.randomize_parameters()
+
+
+class ToTensor(transforms.ToTensor):
+
+    def randomize_parameters(self):
+        pass
+
+
+class Normalize(transforms.Normalize):
+
+    def randomize_parameters(self):
+        pass
+
+
+class ScaleValue(object):
+    """
+    If you set s to 255, the range of inputs is [0-255].
+    """
+
+    def __init__(self, s):
+        self.s = s
+
+    def __call__(self, tensor):
+        tensor *= self.s
+        return tensor
+
+    def randomize_parameters(self):
+        pass
+
+
+class Resize(transforms.Resize):
+    # If size is an int, smaller edge of the image will be matched to this number.
+    def randomize_parameters(self):
+        pass
+
+
+class Scale(transforms.Scale):
+
+    def randomize_parameters(self):
+        pass
+
+
+class CenterCrop(transforms.CenterCrop):
+
+    def randomize_parameters(self):
+        pass
+
+
+class CornerCrop(object):
+
+    def __init__(self,
+                 size,
+                 crop_position=None,
+                 crop_positions=['c', 'tl', 'tr', 'bl', 'br']):
+        self.size = size
+        self.crop_position = crop_position
+        self.crop_positions = crop_positions
+
+        if crop_position is None:
+            self.randomize = True
+        else:
+            self.randomize = False
+        self.randomize_parameters()
+
+    def __call__(self, img):
+        image_width = img.size[0]
+        image_height = img.size[1]
+
+        h, w = (self.size, self.size)
+        if self.crop_position == 'c':
+            i = int(round((image_height - h) / 2.))
+            j = int(round((image_width - w) / 2.))
+        elif self.crop_position == 'tl':
+            i = 0
+            j = 0
+        elif self.crop_position == 'tr':
+            i = 0
+            j = image_width - self.size
+        elif self.crop_position == 'bl':
+            i = image_height - self.size
+            j = 0
+        elif self.crop_position == 'br':
+            i = image_height - self.size
+            j = image_width - self.size
+
+        img = F.crop(img, i, j, h, w)
+
+        return img
+
+    def randomize_parameters(self):
+        if self.randomize:
+            self.crop_position = self.crop_positions[
+                random.randint(0, len(self.crop_positions) - 1)
+            ]
+
+    def __repr__(self):
+        return self.__class__.__name__ + '(size={0}, crop_position={1}, randomize={2})'.format(
+            self.size, self.crop_position, self.randomize)
+
+
+class RandomHorizontalFlip(transforms.RandomHorizontalFlip):
+
+    def __init__(self, p=0.5):
+        super().__init__(p)
+        self.randomize_parameters()
+
+    def __call__(self, img):
+        """
+        Args:
+            img (PIL.Image): Image to be flipped.
+        Returns:
+            PIL.Image: Randomly flipped image.
+        """
+        if self.random_p < self.p:
+            return F.hflip(img)
+        return img
+
+    def randomize_parameters(self):
+        self.random_p = random.random()
+
+
+class RandomVerticalFlip(transforms.RandomVerticalFlip):
+
+    def __init__(self, p=0.5):
+        super().__init__(p)
+        self.randomize_parameters()
+
+    def __call__(self, img):
+        """
+        Args:
+            img (PIL.Image): Image to be flipped.
+        Returns:
+            PIL.Image: Randomly flipped image.
+        """
+        if self.random_p < self.p:
+            return F.vflip(img)
+        return img
+
+    def randomize_parameters(self):
+        self.random_p = random.random()
+
+
+class MultiScaleCornerCrop(object):
+
+    def __init__(self,
+                 size,
+                 scales,
+                 crop_positions=['c', 'tl', 'tr', 'bl', 'br'],
+                 interpolation=Image.BILINEAR):
+        self.size = size
+        self.scales = scales
+        self.interpolation = interpolation
+        self.crop_positions = crop_positions
+
+        self.randomize_parameters()
+
+    def __call__(self, img):
+        short_side = min(img.size[0], img.size[1])
+        crop_size = int(short_side * self.scale)
+        self.corner_crop.size = crop_size
+
+        img = self.corner_crop(img)
+        return img.resize((self.size, self.size), self.interpolation)
+
+    def randomize_parameters(self):
+        self.scale = self.scales[random.randint(0, len(self.scales) - 1)]
+        crop_position = self.crop_positions[
+            random.randint(0, len(self.crop_positions) - 1)
+        ]
+
+        self.corner_crop = CornerCrop(None, crop_position)
+
+    def __repr__(self):
+        return self.__class__.__name__ + '(size={0}, scales={1}, interpolation={2})'.format(
+            self.size, self.scales, self.interpolation)
+
+
+class RandomResizedCrop(transforms.RandomResizedCrop):
+
+    def __init__(self,
+                 size,
+                 scale=(0.08, 1.0),
+                 ratio=(3. / 4., 4. / 3.),
+                 interpolation=Image.BILINEAR):
+        super().__init__(size, scale, ratio, interpolation)
+        self.randomize_parameters()
+
+    def __call__(self, img):
+        if self.randomize:
+            self.random_crop = self.get_params(img, self.scale, self.ratio)
+            self.randomize = False
+
+        i, j, h, w = self.random_crop
+        return F.resized_crop(img, i, j, h, w, self.size, self.interpolation)
+
+    def randomize_parameters(self):
+        self.randomize = True
+
+
+class ColorJitter(transforms.ColorJitter):
+
+    def __init__(self, brightness=0, contrast=0, saturation=0, hue=0):
+        super().__init__(brightness, contrast, saturation, hue)
+        self.randomize_parameters()
+
+    def __call__(self, img):
+        if self.randomize:
+            self.transform = self.get_params(self.brightness, self.contrast,
+                                             self.saturation, self.hue)
+            self.randomize = False
+
+        return self.transform(img)
+
+    def randomize_parameters(self):
+        self.randomize = True
+
+
+class RandomErasing(transforms.RandomErasing):
+    """ Randomly selects a rectangle region in an image and erases its pixels.
+        'Random Erasing Data Augmentation' by Zhong et al.
+        See https://arxiv.org/pdf/1708.04896.pdf
+    """
+
+    def __init__(self, p=0.5, scale=(0.02, 0.33), ratio=(0.3, 3.3), value=0, inplace=False, same_place=True):
+        """
+        same_place (bool):
+            If True, the same spatial place will be erased for every frame of a video.
+            Else, the spatial place to be erased will be decided randomly for each frame.
+        """
+        super().__init__(p, scale, ratio, value, inplace)
+        self.same_place = same_place
+        self.randomize_parameters()
+
+    def __call__(self, img):
+        """
+        Args:
+            img (Tensor): Tensor image of size (C, H, W) to be erased.
+        Returns:
+            img (Tensor): Erased Tensor image.
+        """
+        if self.same_place:
+            if self.random_p < self.p:
+                if self.randomize:
+                    self.erase_params = self.get_params(
+                        img, scale=self.scale, ratio=self.ratio, value=self.value)
+                    self.randomize = False
+
+                x, y, h, w, v = self.erase_params
+                return F.erase(img, x, y, h, w, v, self.inplace)
+            return img
+
+        else:
+            self.random_p = random.random()
+            if self.random_p < self.p:
+                x, y, h, w, v = self.get_params(
+                    img, scale=self.scale, ratio=self.ratio, value=self.value)
+                return F.erase(img, x, y, h, w, v, self.inplace)
+            return img
+
+    def randomize_parameters(self):
+        self.randomize = True
+        self.random_p = random.random()
